requirements.txt
asyncio
websockets
torch
pillow
websocket
torchvision
numpy
python-dotenv
prometheus_client
uvicorn
pydantic-settings
fastapi
redis